src/requirements.txt
appnope==0.1.0
argcomplete==1.8.2
astroid==1.6.5
autopep8==1.4
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
boto==2.49.0
boto3==1.9.14
botocore==1.12.14
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.2
click==6.7
cloudpickle==0.5.5
configparser==3.5.0
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
docutils==0.14
docx2txt==0.6
EbookLib==0.15
enum==0.4.6
enum34==1.1.6
Flask==1.0.2
future==0.16.0
futures==3.2.0
gensim==3.6.0
idna==2.7
imutils==0.5.1
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==2.2.3
mccabe==0.6.1
networkx==2.1
nltk==3.3
numpy==1.15.0
oauthlib==2.1.0
pandas==0.23.4
pathlib2==2.3.2
pbr==4.2.0
pdfminer==20140328
pdfminer.six==20170720
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
playsound==1.2.2
pocketsphinx==0.1.3
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycodestyle==2.4.0
pycryptodome==3.6.6
pygame==1.9.4
Pygments==2.2.0
pylint==1.9.3
pymongo==3.7.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-pptx==0.6.5
python-twitter==3.4.2
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
scandir==1.9.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.7.1
SpeechRecognition==3.6.3
SQLAlchemy==1.2.12
stevedore==1.29.0
subprocess32==3.5.2
TBB==0.1
textract==1.6.1
toolz==0.9.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==1.1.1
